pyproject.toml
[project]
name = "learn-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.3",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.6",
    "langchain-deepseek>=1.0.1",
    "langchain-openai>=1.1.7",
    "python-dotenv>=1.2.1",
]
